Se corrige logica de carrito

+++ src/main/java/com/example/msordenes/application/dto/CarritoItemResponse.java
@@ -1,14 +1,6 @@
 package com.example.msordenes.application.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-import java.math.BigDecimal;
+import lombok.*;
 
 @Getter
 @Setter
@@ -21,6 +13,6 @@ public class CarritoItemResponse {
     private Long idProducto;
     private String nombreProducto;
     private int cantidad;
-    private BigDecimal precioUnitario;
-    private BigDecimal subtotal;
+    private Integer precioUnitario;
+    private Integer subtotal;
 }
