+++ src/main/java/com/example/msordenes/domain/model/Producto.java
@@ -3,14 +3,12 @@ package com.example.msordenes.domain.model;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 public class Producto {
 
     private Long idProducto;
     private String nombre;
-    private BigDecimal precioUnitario;
+    private Integer precioUnitario;
     private String imageUrl;
 }
